Revert "rename to uppercase"
This reverts commit 3448a127ae48198009d8557f78b0974ff168d54c.
