+++ frontend/pages/song/[slug].tsx
@@ -1,4 +1,5 @@
 import { Skeleton } from "@chakra-ui/react";
+import NavigationArrows from "@components/atoms/navigationArrows";
 import SongDetail from "@components/organisms/SongDetail";
 import Layout from "@components/templates/Layout";
 import { NextPage } from "next";
